setup for deployment

+++ frontend/src/pages/Product.jsx
@@ -23,8 +23,10 @@ function Product() {
   }, [fetchProduct, id]);
 
   const handleDelete = async () => {
-    await deleteProduct(id);
-    navigate("/");
+    if (window.confirm("Are you sure you want to delete this product?")) {
+      await deleteProduct(id);
+      navigate("/");
+    }
   };
 
   console.log("THIS IS THE CURRENT PRODUCT", currentProduct);
